+++ zammadoo/resources.py
@@ -3,6 +3,7 @@
 
 import weakref
 from dataclasses import asdict
+from datetime import datetime
 from functools import cached_property, partial
 from typing import (
     TYPE_CHECKING,
@@ -81,6 +82,16 @@ class ResourcesT(Generic[_T_co]):
             return response
 
         return cache.setdefault_by_callback(item, callback)
+
+    def cached_timestamp(self, rid: int) -> Optional[datetime]:
+        return self.cache.timestamp(f"{self.url}/{rid}")
+
+    def delete(self, rid: int) -> None:
+        item = f"{self.url}/{rid}"
+        cache = self.cache
+        self.client.delete(self.endpoint, rid)
+        if item in cache:
+            del cache[item]
 
 
 class CreatableT(ResourcesT[_T_co]):